Added UsageHelpr
Split off console help from UserWrappr.
