+++ Source/FullScreenMario.js
@@ -7108,6 +7108,7 @@ var FullScreenMario;
             "generator": undefined,
             "groups": undefined,
             "events": undefined,
+            "help": undefined,
             "input": undefined,
             "math": undefined,
             "maps": undefined,
