+++ Source/References/UserWrappr-0.2.0.js
@@ -744,9 +744,6 @@ var UserWrappr;
             if (typeof settings.GameStartrConstructor === "undefined") {
                 throw new Error("No GameStartrConstructor given to UserWrappr.");
             }
-            if (typeof settings.helpSettings === "undefined") {
-                throw new Error("No helpSettings given to UserWrappr.");
-            }
             if (typeof settings.globalName === "undefined") {
                 throw new Error("No globalName given to UserWrappr.");
             }
@@ -762,13 +759,11 @@ var UserWrappr;
             this.settings = settings;
             this.GameStartrConstructor = settings.GameStartrConstructor;
             this.globalName = settings.globalName;
-            this.helpSettings = this.settings.helpSettings;
             this.sizes = this.importSizes(settings.sizes);
             this.customs = settings.customs || {};
-            this.gameNameAlias = settings.helpSettings.globalNameAlias || "{GAME}";
             this.gameElementSelector = settings.gameElementSelector || "#game";
             this.gameControlsSelector = settings.gameControlsSelector || "#controls";
-            this.logger = settings.log || console.log.bind(console);
+            this.logger = settings.logger || console.log.bind(console);
             this.isFullScreen = false;
             this.setCurrentSize(this.sizes[settings.sizeDefault]);
             this.allPossibleKeys = settings.allPossibleKeys || UserWrappr.allPossibleKeys;
@@ -829,24 +824,6 @@ var UserWrappr;
          */
         UserWrappr.prototype.getCustoms = function () {
             return this.customs;
-        };
-        /**
-         * @returns The help settings from settings.helpSettings.
-         */
-        UserWrappr.prototype.getHelpSettings = function () {
-            return this.helpSettings;
-        };
-        /**
-         * @returns What the global object is called, such as "window".
-         */
-        UserWrappr.prototype.getGlobalName = function () {
-            return this.globalName;
-        };
-        /**
-         * @returns What to replace with the name of the game in help text.
-         */
-        UserWrappr.prototype.getGameNameAlias = function () {
-            return this.gameNameAlias;
         };
         /**
          * @returns All the keys the user is allowed to pick from in UI controls.
@@ -945,130 +922,6 @@ var UserWrappr;
                 this.GameStarter.container.parentNode.removeChild(this.GameStarter.container);
                 this.resetGameStarter(this.settings, this.customs);
             }
-        };
-        /* Help dialog
-        */
-        /**
-         * Displays the root help menu dialog, which contains all the openings
-         * for each help settings opening.
-         */
-        UserWrappr.prototype.displayHelpMenu = function () {
-            var _this = this;
-            this.helpSettings.openings.forEach(function (opening) { return _this.logHelpText(opening); });
-        };
-        /**
-         * Displays the texts of each help settings options, all surrounded by
-         * instructions on how to focus on a group.
-         */
-        UserWrappr.prototype.displayHelpOptions = function () {
-            var _this = this;
-            this.logHelpText([
-                ("To focus on a group, enter %c" + this.globalName + ".UserWrapper.displayHelpOption(\"<group-name>\");%c"),
-                "code"
-            ]);
-            Object.keys(this.helpSettings.options).forEach(function (key) { return _this.displayHelpGroupSummary(key); });
-            this.logHelpText([
-                ("\nTo focus on a group, enter %c" + this.globalName + ".UserWrapper.displayHelpOption(\"<group-name>\");%c"),
-                "code"
-            ]);
-        };
-        /**
-         * Displays the summary for a help group of the given optionName.
-         *
-         * @param optionName   The help group to display the summary of.
-         */
-        UserWrappr.prototype.displayHelpGroupSummary = function (optionName) {
-            var actions = this.helpSettings.options[optionName], action, maxTitleLength = 0, i;
-            this.logger("\r\n%c" + optionName, UserWrappr.styles.head);
-            for (i = 0; i < actions.length; i += 1) {
-                maxTitleLength = Math.max(maxTitleLength, this.filterHelpText(actions[i].title).length);
-            }
-            for (i = 0; i < actions.length; i += 1) {
-                action = actions[i];
-                this.logger("%c" + this.padTextRight(this.filterHelpText(action.title), maxTitleLength) + "%c  // " + action.description, UserWrappr.styles.code, UserWrappr.styles.comment);
-            }
-        };
-        /**
-         * Displays the full information on a help group of the given optionName.
-         *
-         * @param optionName   The help group to display the information of.
-         */
-        UserWrappr.prototype.displayHelpOption = function (optionName) {
-            var actions = this.helpSettings.options[optionName], action, example, maxExampleLength, i, j;
-            this.logHelpText([("\r\n\r\n%c" + optionName + "\r\n-------\r\n\r\n"), "head"]);
-            for (i = 0; i < actions.length; i += 1) {
-                action = actions[i];
-                maxExampleLength = 0;
-                this.logHelpText([
-                    ("%c" + action.title + "%c  ---  " + action.description),
-                    "head",
-                    "italic"
-                ]);
-                if (action.usage) {
-                    this.logHelpText([
-                        ("%cUsage: %c" + action.usage),
-                        "comment",
-                        "code"
-                    ]);
-                }
-                if (action.examples) {
-                    for (j = 0; j < action.examples.length; j += 1) {
-                        example = action.examples[j];
-                        this.logger("\r\n");
-                        this.logHelpText([("%c// " + example.comment), "comment"]);
-                        this.logHelpText([
-                            ("%c" + this.padTextRight(this.filterHelpText(example.code), maxExampleLength)),
-                            "code"
-                        ]);
-                    }
-                }
-                this.logger("\r\n");
-            }
-        };
-        /**
-         * Logs a bit of help text, filtered by this.filterHelpText, with ordered styles
-         * from `UserWrappr.styles` keyed by name.
-         *
-         * @param text   The text to be filtered and logged.
-         * @remarks See https://getfirebug.com/wiki/index.php/Console.log for "%c" usage.
-         */
-        UserWrappr.prototype.logHelpText = function (line) {
-            if (typeof line === "string") {
-                return this.logHelpText([line]);
-            }
-            var message = line[0], styles = line
-                .slice(1)
-                .filter(function (style) { return UserWrappr.styles.hasOwnProperty(style); })
-                .map(function (style) { return UserWrappr.styles[style]; });
-            // A last blank "" style  allows the last "%c" in the message to reset text styles
-            this.logger.apply(this, [this.filterHelpText(message)].concat(styles, [""]));
-        };
-        /**
-         * Filters a span of help text to replace the game name with its alias. If "%c" isn't
-         * in the text, it's added at the end.
-         *
-         * @param text The text to filter.
-         * @returns The text, with `this.gameNameAlias` replaced by globalName.
-         */
-        UserWrappr.prototype.filterHelpText = function (textRaw) {
-            if (textRaw.constructor === Array) {
-                return this.filterHelpText(textRaw[0]);
-            }
-            return textRaw.replace(new RegExp(this.gameNameAlias, "g"), this.globalName);
-        };
-        /**
-         * Ensures a bit of text is of least a certain length.
-         *
-         * @param text   The text to pad.
-         * @param length   How wide the text must be, at minimum.
-         * @returns The text with spaces padded to the right.
-         */
-        UserWrappr.prototype.padTextRight = function (text, length) {
-            var diff = 1 + length - text.length;
-            if (diff <= 0) {
-                return text;
-            }
-            return text + Array.call(Array, diff).join(" ");
         };
         /* Devices
         */
@@ -1266,16 +1119,6 @@ var UserWrappr;
             "n", "o", "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z",
             "up", "right", "down", "left", "space", "shift", "ctrl"
         ];
-        /**
-         * Styles for fancy text in console help messages.
-         */
-        UserWrappr.styles = {
-            "code": "color: #000077; font-weight: bold; font-family: Consolas, Courier New, monospace;",
-            "comment": "color: #497749; font-style: italic;",
-            "head": "font-weight: bold; font-size: 117%;",
-            "italic": "font-style: italic;",
-            "none": ""
-        };
         return UserWrappr;
     })();
     UserWrappr_1.UserWrappr = UserWrappr;
