+++ Source/References/UserWrappr-0.2.0.ts
@@ -70,104 +70,6 @@ declare module UserWrappr {
     }
 
     /**
-     * A single line of text to write to a console. If an Array, the first String is the 
-     * message, and any others are aliases of UserWrappr styles to apply.
-     */
-    export type IHelpLine = string | string[];
-
-    /**
-     * Descriptions of help settings to display in the console.
-     */
-    export interface IUIHelpSettings {
-        /**
-         * An alias to replace with the IGameStartr's globalName.
-         */
-        globalNameAlias: string;
-
-        /**
-         * Lines to display immediately upon starting.
-         */
-        openings: IHelpLine[];
-
-        /**
-         * Descriptions of APIs users may use, along with sample code.
-         */
-        options: {
-            [i: string]: IHelpOption[];
-        };
-    }
-
-    /**
-     * Descriptions of APIs users may use, along with sample code.
-     */
-    export interface IHelpOption {
-        /**
-         * A label for the API to research it by.
-         */
-        title: string;
-
-        /**
-         * A common description of the API.
-         */
-        description: string;
-
-        /**
-         * Code sample for usage of the API.
-         */
-        usage?: string;
-
-        /**
-         * API code samples with explanations.
-         */
-        examples?: IHelpExample[];
-    }
-
-    /**
-     * Code sample for an API with an explanation.
-     */
-    export interface IHelpExample {
-        /**
-         * An API code sample.
-         */
-        code: IHelpLine;
-
-        /**
-         * An explanation for the API code sample.
-         */
-        comment: IHelpLine;
-    }
-
-    /**
-     * Styles that may be applied to console help text.
-     */
-    export interface ITextStyles {
-        /**
-         * Style for headers.
-         */
-        head: string;
-
-        /**
-         * Style for code.
-         */
-        code: string;
-
-        /**
-         * Style for code comments.
-         */
-        comment: string;
-
-        /**
-         * Style for italicized text.
-         */
-        italic: string;
-
-        /**
-         * No styles at all (plain text).
-         */
-        none: string;
-    }
-
-    /**
      * How wide and tall an IUserWrappr's contained IGameStartr should be sized.
      */
     export interface ISizeSummary {
@@ -203,11 +105,6 @@ declare module UserWrappr {
      * Settings to initialize a new IUserWrappr.
      */
     export interface IUserWrapprSettings {
-        /**
-         * Descriptions of help settings to display in the console.
-         */
-        helpSettings: IUIHelpSettings;
-
         /**
          * What the global object is called, such as "window".
          */
@@ -246,7 +143,7 @@ declare module UserWrappr {
         /**
          * A utility Function to log messages, commonly console.log.
          */
-        log?: (...args: any[]) => void;
+        logger?: (...args: any[]) => void;
 
         /**
          * Custom arguments to be passed to the IGameStartr's modules.
@@ -302,21 +199,6 @@ declare module UserWrappr {
          * @returns The customs used to construct the IGameStartr.
          */
         getCustoms(): IGameStartrCustoms;
-
-        /**
-         * @returns The help settings from settings.helpSettings.
-         */
-        getHelpSettings(): IUIHelpSettings;
-
-        /**
-         * @returns What the global object is called, such as "window".
-         */
-        getGlobalName(): string;
-
-        /**
-         * @returns What to replace with the name of the game in help text.
-         */
-        getGameNameAlias(): string;
 
         /**
          * @returns All the keys the user is allowed to pick from in UI controls.
@@ -382,32 +264,6 @@ declare module UserWrappr {
          *             known info, or a container of settings.
          */
         setCurrentSize(size: string | ISizeSummary): void;
-
-        /**
-         * Displays the root help menu dialog, which contains all the openings
-         * for each help settings opening.
-         */
-        displayHelpMenu(): void;
-
-        /**
-         * Displays the texts of each help settings options, all surrounded by
-         * instructions on how to focus on a group.
-         */
-        displayHelpOptions(): void;
-
-        /**
-         * Displays the summary for a help group of the given optionName.
-         * 
-         * @param optionName   The help group to display the summary of.
-         */
-        displayHelpGroupSummary(optionName: string): void;
-
-        /**
-         * Displays the full information on a help group of the given optionName.
-         * 
-         * @param optionName   The help group to display the information of.
-         */
-        displayHelpOption(optionName: string): void;
     }
 
     /**
@@ -1689,17 +1545,6 @@ module UserWrappr {
         ];
 
         /**
-         * Styles for fancy text in console help messages.
-         */
-        private static styles: ITextStyles = {
-            "code": "color: #000077; font-weight: bold; font-family: Consolas, Courier New, monospace;",
-            "comment": "color: #497749; font-style: italic;",
-            "head": "font-weight: bold; font-size: 117%;",
-            "italic": "font-style: italic;",
-            "none": ""
-        };
-
-        /**
          * The GameStartr implementation this is wrapping around, such as
          * FullScreenMario or FullScreenPokemon.
          */
@@ -1727,21 +1572,10 @@ module UserWrappr {
         private customs: any;
 
         /**
-         * Help settings specifically for the user interface, obtained from
-         * settings.helpSettings.
-         */
-        private helpSettings: IUIHelpSettings;
-
-        /**
          * What the global object is called (typically "window" for browser 
          * environments and "global" for node-style environments).
          */
         private globalName: string;
-
-        /**
-         * What to replace with the name of the game in help text.
-         */
-        private gameNameAlias: string;
 
         /**
          * All the keys the user is allowed to pick from as key bindings.
@@ -1836,9 +1670,6 @@ module UserWrappr {
             if (typeof settings.GameStartrConstructor === "undefined") {
                 throw new Error("No GameStartrConstructor given to UserWrappr.");
             }
-            if (typeof settings.helpSettings === "undefined") {
-                throw new Error("No helpSettings given to UserWrappr.");
-            }
             if (typeof settings.globalName === "undefined") {
                 throw new Error("No globalName given to UserWrappr.");
             }
@@ -1855,15 +1686,13 @@ module UserWrappr {
             this.settings = settings;
             this.GameStartrConstructor = settings.GameStartrConstructor;
             this.globalName = settings.globalName;
-            this.helpSettings = this.settings.helpSettings;
 
             this.sizes = this.importSizes(settings.sizes);
 
             this.customs = settings.customs || {};
-            this.gameNameAlias = settings.helpSettings.globalNameAlias || "{GAME}";
             this.gameElementSelector = settings.gameElementSelector || "#game";
             this.gameControlsSelector = settings.gameControlsSelector || "#controls";
-            this.logger = settings.log || console.log.bind(console);
+            this.logger = settings.logger || console.log.bind(console);
 
             this.isFullScreen = false;
             this.setCurrentSize(this.sizes[settings.sizeDefault]);
@@ -1941,27 +1770,6 @@ module UserWrappr {
          */
         getCustoms(): IGameStartrCustoms {
             return this.customs;
-        }
-
-        /**
-         * @returns The help settings from settings.helpSettings.
-         */
-        getHelpSettings(): IUIHelpSettings {
-            return this.helpSettings;
-        }
-
-        /**
-         * @returns What the global object is called, such as "window".
-         */
-        getGlobalName(): string {
-            return this.globalName;
-        }
-
-        /**
-         * @returns What to replace with the name of the game in help text.
-         */
-        getGameNameAlias(): string {
-            return this.gameNameAlias;
         }
 
         /**
@@ -2079,167 +1887,6 @@ module UserWrappr {
             }
         }
 
-
-        /* Help dialog
-        */
-
-        /**
-         * Displays the root help menu dialog, which contains all the openings
-         * for each help settings opening.
-         */
-        displayHelpMenu(): void {
-            this.helpSettings.openings.forEach(
-                (opening: string[]): void => this.logHelpText(opening));
-        }
-
-        /**
-         * Displays the texts of each help settings options, all surrounded by
-         * instructions on how to focus on a group.
-         */
-        displayHelpOptions(): void {
-            this.logHelpText([
-                `To focus on a group, enter %c${this.globalName}.UserWrapper.displayHelpOption("<group-name>");%c`,
-                "code"
-            ]);
-
-            Object.keys(this.helpSettings.options).forEach(
-                (key: string): void => this.displayHelpGroupSummary(key));
-
-            this.logHelpText([
-                `\nTo focus on a group, enter %c${this.globalName}.UserWrapper.displayHelpOption("<group-name>");%c`,
-                "code"
-            ]);
-        }
-
-        /**
-         * Displays the summary for a help group of the given optionName.
-         * 
-         * @param optionName   The help group to display the summary of.
-         */
-        displayHelpGroupSummary(optionName: string): void {
-            var actions: IHelpOption[] = this.helpSettings.options[optionName],
-                action: IHelpOption,
-                maxTitleLength: number = 0,
-                i: number;
-
-            this.logger(`\r\n%c${optionName}`, UserWrappr.styles.head);
-
-            for (i = 0; i < actions.length; i += 1) {
-                maxTitleLength = Math.max(maxTitleLength, this.filterHelpText(actions[i].title).length);
-            }
-
-            for (i = 0; i < actions.length; i += 1) {
-                action = actions[i];
-                this.logger(
-                    `%c${this.padTextRight(this.filterHelpText(action.title), maxTitleLength)}%c  // ${action.description}`,
-                    UserWrappr.styles.code,
-                    UserWrappr.styles.comment);
-            }
-        }
-
-        /**
-         * Displays the full information on a help group of the given optionName.
-         * 
-         * @param optionName   The help group to display the information of.
-         */
-        displayHelpOption(optionName: string): void {
-            var actions: IHelpOption[] = this.helpSettings.options[optionName],
-                action: IHelpOption,
-                example: IHelpExample,
-                maxExampleLength: number,
-                i: number,
-                j: number;
-
-            this.logHelpText([`\r\n\r\n%c${optionName}\r\n-------\r\n\r\n`, "head"]);
-
-            for (i = 0; i < actions.length; i += 1) {
-                action = actions[i];
-                maxExampleLength = 0;
-
-                this.logHelpText([
-                    `%c${action.title}%c  ---  ${action.description}`,
-                    "head",
-                    "italic"
-                ]);
-
-                if (action.usage) {
-                    this.logHelpText([
-                        `%cUsage: %c${action.usage}`,
-                        "comment",
-                        "code"
-                    ]);
-                }
-
-                if (action.examples) {
-                    for (j = 0; j < action.examples.length; j += 1) {
-                        example = action.examples[j];
-
-                        this.logger("\r\n");
-                        this.logHelpText([`%c// ${example.comment}`, "comment"]);
-                        this.logHelpText([
-                            `%c${this.padTextRight(this.filterHelpText(example.code), maxExampleLength)}`,
-                            "code"
-                        ]);
-                    }
-                }
-
-                this.logger("\r\n");
-            }
-        }
-
-        /**
-         * Logs a bit of help text, filtered by this.filterHelpText, with ordered styles
-         * from `UserWrappr.styles` keyed by name.
-         * 
-         * @param text   The text to be filtered and logged.
-         * @remarks See https://getfirebug.com/wiki/index.php/Console.log for "%c" usage.
-         */
-        private logHelpText(line: IHelpLine): void {
-            if (typeof line === "string") {
-                return this.logHelpText([line]);
-            }
-
-            var message: string = line[0],
-                styles: string[] = (<string[]>line)
-                    .slice(1)
-                    .filter((style: string): boolean => UserWrappr.styles.hasOwnProperty(style))
-                    .map((style: string): string => UserWrappr.styles[style]);
-
-            // A last blank "" style  allows the last "%c" in the message to reset text styles
-            this.logger(this.filterHelpText(message), ...styles, "");
-        }
-
-        /**
-         * Filters a span of help text to replace the game name with its alias. If "%c" isn't
-         * in the text, it's added at the end.
-         * 
-         * @param text The text to filter.
-         * @returns The text, with `this.gameNameAlias` replaced by globalName.
-         */
-        private filterHelpText(textRaw: IHelpLine): string {
-            if (textRaw.constructor === Array) {
-                return this.filterHelpText(textRaw[0]);
-            }
-
-            return (<string>textRaw).replace(new RegExp(this.gameNameAlias, "g"), this.globalName);
-        }
-
-        /**
-         * Ensures a bit of text is of least a certain length.
-         * 
-         * @param text   The text to pad.
-         * @param length   How wide the text must be, at minimum.
-         * @returns The text with spaces padded to the right.
-         */
-        private padTextRight(text: string, length: number): string {
-            var diff: number = 1 + length - text.length;
-
-            if (diff <= 0) {
-                return text;
-            }
-
-            return text + Array.call(Array, diff).join(" ");
-        }
 
         /* Devices
         */
