+++ Source/index.html
@@ -92,6 +92,7 @@
     <script src="References/ThingHittr-0.2.0.js" type="text/javascript"></script>
     <script src="References/TimeHandlr-0.2.0.js" type="text/javascript"></script>
     <script src="References/TouchPassr-0.2.0.js" type="text/javascript"></script>
+    <script src="References/UsageHelpr-0.2.0.js" type="text/javascript"></script>
     <script src="References/UserWrappr-0.2.0.js" type="text/javascript"></script>
     <script src="References/WorldSeedr-0.2.0.js" type="text/javascript"></script>
     <script src="References/js_beautify.js" type="text/javascript"></script>
@@ -103,6 +104,7 @@
     <script src="settings/generator.js" type="text/javascript"></script>
     <script src="settings/groups.js" type="text/javascript"></script>
     <script src="settings/events.js" type="text/javascript"></script>
+    <script src="settings/help.js" type="text/javascript"></script>
     <script src="settings/input.js" type="text/javascript"></script>
     <script src="settings/maps.js" type="text/javascript"></script>
     <script src="settings/maps/1-1.js" type="text/javascript"></script>
