+++ Source/index.js
@@ -10,5 +10,7 @@ document.onreadystatechange = function (event) {
             "GameStartrConstructor": FullScreenMario.FullScreenMario
         }, FullScreenMario.FullScreenMario.settings.ui, true));
 
-    console.log("It took " + (Date.now() - time) + " milliseconds to start."), UserWrapper.displayHelpMenu()
+    console.log("It took " + (Date.now() - time) + " milliseconds to start.");
+    
+    UserWrapper.GameStarter.UsageHelper.displayHelpMenu();
 };
