+++ Tests/tests.html
@@ -35,6 +35,7 @@
     <script src="../Source/References/ThingHittr-0.2.0.js" type="text/javascript"></script>
     <script src="../Source/References/TimeHandlr-0.2.0.js" type="text/javascript"></script>
     <script src="../Source/References/TouchPassr-0.2.0.js" type="text/javascript"></script>
+    <script src="../Source/References/UsageHelpr-0.2.0.js" type="text/javascript"></script>
     <script src="../Source/References/UserWrappr-0.2.0.js" type="text/javascript"></script>
     <script src="../Source/References/WorldSeedr-0.2.0.js" type="text/javascript"></script>
     <script src="../Source/References/js_beautify.js" type="text/javascript"></script>
@@ -45,6 +46,7 @@
     <script src="../Source/settings/generator.js"></script>
     <script src="../Source/settings/groups.js"></script>
     <script src="../Source/settings/events.js"></script>
+    <script src="../Source/settings/help.js"></script>
     <script src="../Source/settings/input.js"></script>
     <script src="../Source/settings/maps.js"></script>
     <script src="../Source/settings/maps/1-1.js" type="text/javascript"></script>
